requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-generativeai
python-dotenv
beautifulsoup4
playwright